pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "redis[hiredis] (>=5.2.1,<6.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "celery (>=5.5.1,<6.0.0)",
    "poetry-plugin-export (>=1.9.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "dj-database-url (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
